chore: Move main `MetadataRetrieverModule` to `modules` folder
