+++ android/src/main/java/com/cyanchill/missingcore/metadataretriever/MetadataRetrieverPackage.kt
@@ -7,6 +7,7 @@ import com.facebook.react.module.model.ReactModuleInfoProvider
 import com.facebook.react.module.model.ReactModuleInfo
 import java.util.HashMap
 
+import com.cyanchill.missingcore.metadataretriever.modules.MetadataRetrieverModule
 
 class MetadataRetrieverPackage : TurboReactPackage() {
   override fun getModule(name: String, reactContext: ReactApplicationContext): NativeModule? {
